+++ backend/Models/M04/IpqcFirstPiece.cs
@@ -22,13 +22,13 @@ public class IpqcFirstPiece
     public long WorkOrderId { get; set; }
 
     /// <summary>关联工序</summary>
-    public int ProcessId { get; set; }
+    public long ProcessId { get; set; }
 
     /// <summary>关联设备</summary>
-    public int EquipmentId { get; set; }
+    public long EquipmentId { get; set; }
 
     /// <summary>操作员</summary>
-    public int OperatorId { get; set; }
+    public long OperatorId { get; set; }
 
     /// <summary>班次：早班/中班/晚班</summary>
     [MaxLength(20)]
@@ -48,7 +48,7 @@ public class IpqcFirstPiece
     public bool AllowedToProduce { get; set; }
 
     /// <summary>检验员</summary>
-    public int? InspectorId { get; set; }
+    public long? InspectorId { get; set; }
 
     /// <summary>检验时间</summary>
     public DateTime? CheckedAt { get; set; }
